Fix flake8 errors (#51)
