+++ .circleci/config.yml
@@ -43,7 +43,7 @@ jobs:
           command: sudo pip install pytest pytest-cov codecov
       - run:
           name: Run tests
-          command: pytest tst.py --cov=./
+          command: pytest test_all.py --cov=./
       - store_artifacts:
           path: test-reports
           destination: test-reports
